RH tab and A JBoss project box, and mobile adjustments.
Made css changes to include the “A JBoss project” box to the left of
the Tab. This included changing the way we’ve coded the css for the
subproject pages. I also made adjustments to the overall padding around
he pages to remove redundant backgrounds when the page is resized to
mobile. I also moved up the RH tab related css to the top of the page
and added comments so it’s easier to track.